Add WanderingTraderSerializer
